Working with Enzyme

The Julia ecosystem has had (infamously) a number of automatic differentiation packages, each with their own tradeoffs. Wren adopts a partly opinionated approach to AD. All (sensible) AD backends are supported via DifferentiationInterface.jl, so the user can choose their preferred backend. However, at the same time, Wren also tries very hard to make it possible for you to use reverse-mode Enzyme as the backend.

The developer(s) of Wren are not affiliated with any AD backend and this opinion is not rooted in any politics or personal disagreement. It is purely based on the fact that reverse Enzyme is the fastest AD backend across a wide range of problems (see, e.g., the benchmarks on Turing.jl). Thus, we consider it a disservice to choose any other backend as the default, as that simply leaves performance on the table.

At the same time, we recognise that Enzyme can be difficult to use and still has a number of edge cases. Wren does its best to make this experience as seamless as possible for users. However, it is greatly appreciated if you help us report cases where Enzyme fails to work, so that we can improve the experience for everyone. We can (usually) help with minimising bugs and filing issues upstream. Thank you in advance!

Runtime activity

Wren does not generally require you to enable runtime activity when using Enzyme. Thus, you can simply use AutoEnzyme(; mode=Reverse) as the adtype when e.g. sampling with NUTS, or constructing a VecModel.

However, depending on how the model is written, sometimes it is necessary for Enzyme to use runtime activity. Usually this situation causes an error to be thrown. However, Wren will, by default, detect such situations and automatically enable runtime activity for you, issuing a warning in the process. Note that even with runtime activity enabled, Enzyme still tends to be the fastest AD backend, so you do not really need to be concerned about this message.

This behaviour can be controlled via the function WrenCore.set_auto_runtime_activity!:

WrenCore.set_auto_runtime_activity! — Function

WrenCore.set_auto_runtime_activity!(val::Union{Bool,Nothing})

Set Wren's automatic runtime activity behaviour.

When using Enzyme as the AD backend, depending on how the model is constructed, it may be necessary to enable runtime activity. This can often allow more expressive models to be differentiated, but comes at a performance cost.

Wren recommends using reverse-mode Enzyme, i.e., AutoEnzyme(; mode=Reverse) as the AD backend, which does not enable runtime activity. By default, Wren will automatically enable runtime activity if Enzyme requires it, but will issue an info message.

If you would rather have Wren raise an error, set

WrenCore.set_auto_runtime_activity!(false)

Conversely, if you are happy with this and want to suppress the info message, set

WrenCore.set_auto_runtime_activity!(true)

To restore the default behaviour (automatic enabling of runtime activity but with an info message), set

WrenCore.set_auto_runtime_activity!(nothing)

When exactly is runtime activity required?

The easiest situation where runtime activity will be required is when differentiable and non-differentiable data are bundled into the same struct. Often there is not much you can do as a user to avoid this situation, which is why we intentionally catch the error for you.

Differentiable data, in the context of Wren, refers to the actual parameters of the model (i.e., the things we want to know gradients with respect to). For example, if you have x ~ MvNormal(...), then x contains differentiable data during model evaluation. Non-differentiable data usually means constants that are supplied to the model.

Consider, for example, the following model:

using Wren, Distributions, Enzyme@wren function f(sigma, y)    m ~ MvNormal(zeros(4), I)    y ~ MvNormal(m, Diagonal(sigma))end

f (generic function with 1 method)

Here, m is differentiable data, whereas sigma is non-differentiable data. Mixing these two things in the same struct (i.e., MvNormal) will trigger Enzyme's runtime activity requirement.

However, at the same time, it's clearly not possible to rewrite the model to avoid this. If the offending struct is only used as an observation distribution, then you can manually write out the log-likelihood and use @addloglikelihood and thus avoid making the struct, but this is less expressive and will make your model less usable.

To be precise...

The need for runtime activity in the above case can also be overcome by creating a copy of sigma inside the model. In fact, because of certain other Enzyme-related requirements, Wren does actually copy all model arguments during evaluation, so this model as written above does not require runtime activity:

sigma = ones(4);y = randn(4)WrenCore.set_auto_runtime_activity!(false)# Should error if it needs runtime activity; but doesn'tVecModel(f(sigma, y), AutoEnzyme(; mode=Reverse))

VecModel f
├─ adtype => AutoEnzyme(mode=EnzymeCore.ReverseMode{false, false, false, EnzymeCore.FFIABI, false, false}())
├─ dimension => 8
├─ parameter space => UnconstrainedSpace()
└─ Model f
   ├─ arguments
   │  ├─ sigma => [1.0, 1.0, 1.0, 1.0]
   │  └─ y => [-1.55855, -0.0443527, 1.54078, -1.64652]
   ├─ conditioned (none)
   └─ expressions
        m ~ MvNormal(zeros(4), I)
        y ~ MvNormal(m, Diagonal(sigma))

However, note that creating a copy comes with its own performance cost so this is not a free solution.

Copying model arguments on evaluation

Warning

You should not need to read this unless you are actively developing Wren.

When originally written, Wren did not copy any model arguments during model evaluation (it would copy them when creating the mode, but not when evaluating it, since the latter is usually done many more times). This had to be changed because it could silently cause incorrect results with Enzyme.

The failure mode is when differentiable data is written into a model argument, and then read out again. Consider the following model, which is only slightly modified from the previous section's to get rid of sigma:

@wren function f(m, y)    m[1:4] ~ MvNormal(zeros(4), I)    @observe y[1:4] ~ MvNormal(m[1:4], I)end

The astute reader will notice that this is actually a batched for-loop in disguise:

@wren function f(m, y)    @batch for i in 1:4        m[i] ~ Normal(0, 1)        @observe y[i] ~ Normal(m[i], 1)    endend

Anyway, Wren emits a model evaluation function that conceptually looks like:

function f_eval(params, m, y)    lp = 0.0    m[1:4] = params[1:4]    lp += logpdf(MvNormal(zeros(4), I), m[1:4])    lp += logpdf(MvNormal(m[1:4], I), y[1:4])    return lpend

Notice that here params is differentiable, but m and y are not. Wren marks the latter two (the model arguments) with DI.Constant, which then translates into Enzyme.Const, so the invocation of AD is morally equivalent to

Enzyme.gradient(Reverse, f_eval, params, Const(m), Const(y))

Since the log-likelihood term (y[1:4] ~ ...) depends on m[1:4], which in turn is derived from params[1:4], the log-likelihood should contribute to the gradient with respect to params.

However, this is unfortunately lost when we write the differentiable data from params into the non-differentiable m:

m[1:4] = params[1:4]

Since m was declared as Const, it doesn't have a 'shadow': in other words, it doesn't have a corresponding array to store its gradients in. So when we reach the log-likelihood term, Enzyme has no way of 'tracking' the gradient of the log-likelihood with respect to params.

This behaviour is intentional in Enzyme and is a result of the Const marking. There isn't a simple way to work around this with DifferentiationInterface, so Wren currently works around this by copying all model arguments during evaluation.

In principle, this can be somewhat relaxed by inspecting the model AST to see how many times each model argument is used. In general, if a model argument is only used once, then it can't fall into the trap above, and we can safely elide the copy. This optimisation has not been implemented yet, though.

WrenCore.set_auto_runtime_activity!(val::Union{Bool,Nothing})
Set Wren's automatic runtime activity behaviour.

VecModel( model::Model, adtype::Union{ADTypes.AbstractADType,Nothing}; lp_function=logjoint_transformed, latent_source=FromPrior(), parameter_space=UnconstrainedSpace() )
A vectorised form of model, which is designed to take a single vector of parameters as input, and returns a scalar log-density.

lp(vecmodel::VecModel, params::AbstractVector)
Calculate the log-density of vecmodel at params.

WrenCore.set_auto_runtime_activity!(val::Union{Bool,Nothing})
Set Wren's automatic runtime activity behaviour.

Normal(μ, σ) Normal(μ) Normal()
Normal distribution with mean μ (default 0.0) and standard deviation σ (default 1.0).